+++ src/RESTable/Resources/Operations/IAsyncDeleter.cs
@@ -14,6 +14,6 @@ namespace RESTable.Resources.Operations
         /// The delete method for this IDeleter instance. Defines the Delete
         /// operation for a given resource.
         /// </summary>
-        ValueTask<int> DeleteAsync(IRequest<T> request, CancellationToken cancellationToken);
+        ValueTask<long> DeleteAsync(IRequest<T> request, CancellationToken cancellationToken);
     }
 }
